Moved simple-types tests into existing files for the two main operations
